+++ src/main.js
@@ -16,19 +16,20 @@
     11. Palette (color palette)
     12. Helper (overlays)
     13. Tool
-    14. Symmetry
-    15. Voxelizer
-    16. Generator
-    17. Bakery
-    18. Snapshot
-    19. Memory
-    20. Project
-    21. UserInterface
-    22. UserInterfaceAdvanced
-    23. Preferences
-    24. Events
-    25. Websocket
-    26. Utils
+    14. Tool Bakery
+    15. Symmetry
+    16. Voxelizer
+    17. Generator
+    18. Bakery
+    19. Snapshot
+    20. Memory
+    21. Project
+    22. UserInterface
+    23. UserInterfaceAdvanced
+    24. Preferences
+    25. Events
+    26. Websocket
+    27. Utils
 */
 //data:image/svg+xml,<svg xmlns=%22http://www.w3.org/2000/svg%22 viewBox=%220 0 100 90%22><text x=%220.18em%22 y=%221.1em%22 font-size=%2260%22 style="filter: grayscale(); opacity: 0.15;">❌</text></svg>
 
@@ -67,6 +68,7 @@ const COL_AXIS_Z = '#2F81DF';
 const COL_AXIS_Z_RGB = BABYLON.Color3.FromHexString(COL_AXIS_Z);
 const COL_AXIS_Z_RGBA = BABYLON.Color4.FromHexString(COL_AXIS_Z + 'FF');
 const COL_RED = '#FF0000';
+const COL_CLEAR_RGBA = new BABYLON.Color4(0, 0, 0, 0);
 
 const PI2 = Math.PI * 2;
 const PIH = Math.PI / 2;
@@ -76,7 +78,7 @@ const MAX_VOXELS = 512000;
 const MAX_VOXELS_DRAW = 64000;
 const WORKPLANE_SIZE = 120;
 const WORKPLANE_VISIBILITY = 0.2;
-const DUPLICATE_POS = new BABYLON.Vector3(-1000000, -1000000, -1000000);
+const DUPLICATE_POS = new BABYLON.Vector3(-2000000, -2000000, -2000000);
 const DEF_BGCOLOR = document.getElementById('input-env-background').value.toUpperCase();
 const DEF_LIGHTCOLOR = document.getElementById('input-light-color').value.toUpperCase();
 
@@ -127,7 +129,8 @@ camera.init(scene);
 const sceneAxisView = createAxisViewScene(engine, scene);
 const light = new Light(scene);
 
-const renderPickTarget = new BABYLON.RenderTargetTexture('picktex', { useSRGBBuffer: false, delayAllocation: true, noColorAttachment: true, width: engine.getRenderWidth(), height: engine.getRenderHeight() }, scene, true);
+const renderPickTarget = new BABYLON.RenderTargetTexture('pickingTexure', { useSRGBBuffer: false, width: engine.getRenderWidth(), height: engine.getRenderHeight() }, scene, false, undefined, BABYLON.Constants.TEXTURETYPE_UNSIGNED_INT, false, BABYLON.Constants.TEXTURE_NEAREST_NEAREST);
+renderPickTarget.clearColor = COL_CLEAR_RGBA;
 scene.customRenderTargets.push(renderPickTarget);
 
 const ui = new UserInterface(scene);
@@ -146,6 +149,7 @@ const helper = new Helper(scene, sceneAxisView);
 const symmetry = new Symmetry();
 
 const tool = new Tool(scene);
+const toolBakery = new ToolBakery(scene);
 
 const project = new Project(scene);
 const snapshot = new Snapshot(scene);
@@ -184,31 +188,30 @@ scene.registerAfterRender(() => {
             sceneAxisView.activeCamera.beta = scene.activeCamera.beta;
         }
         
-        if (MODE == 0 && !tool.isMouseDown)
+        if (MODE == 0 && !tool.isMouseDown) {
             camera.lastPos = [ camera.camera0.alpha, camera.camera0.beta ];
+
+            if (duplicateFlag == 1) {
+                builder.removeDuplicates();
+                duplicateFlag = 0;
+            }
+        }
         
         if (scene.activeCamera.mode == BABYLON.Camera.ORTHOGRAPHIC_CAMERA)
             camera.setOrthoMode();
-
-        if (duplicateFlag == 1) {
-            builder.removeDuplicates();
-            duplicateFlag = 0;
-        }
 
         ui.updateStatus();
     }
 });
 
 renderPickTarget.onBeforeRender = () => {
-    if (isRendering)
-        if (MODE == 0 && tool.name !== 'camera')
-            builder.mesh.thinInstanceSetBuffer("color", builder.rttColors, 4, true);
+    if (isRendering && MODE == 0 && tool.name !== 'camera')
+        builder.mesh.thinInstanceSetBuffer("color", builder.rttColors, 4, true);
 }
 
 renderPickTarget.onAfterRender = () => {
-    if (isRendering)
-        if (MODE == 0 && tool.name !== 'camera')
-            builder.mesh.thinInstanceSetBuffer("color", builder.bufferColors, 4, true);
+    if (isRendering && MODE == 0 && tool.name !== 'camera')
+        builder.mesh.thinInstanceSetBuffer("color", builder.bufferColors, 4, true);
 }
 
 
@@ -218,7 +221,7 @@ renderPickTarget.onAfterRender = () => {
 
 function createScene(engine) {
     const scene = new BABYLON.Scene(engine);
-    scene.clearColor = new BABYLON.Color4(0, 0, 0, 0);
+    scene.clearColor = COL_CLEAR_RGBA;
     scene.autoClear = false;
     scene.autoClearDepthAndStencil = false;
     scene.blockMaterialDirtyMechanism = true;
@@ -251,9 +254,11 @@ function createScene(engine) {
 
 function createAxisViewScene(engine, mainScene) {
     const scene = new BABYLON.Scene(engine);
+    scene.clearColor = COL_CLEAR_RGBA;
     scene.autoClear = false;
-    scene.autoClearDepthAndStencil = true;
+    scene.autoClearDepthAndStencil = false;
     scene.blockMaterialDirtyMechanism = true;
+    scene.collisionsEnabled = false;
     scene.useRightHandedSystem = true;
 
     const ambient = new BABYLON.HemisphericLight("ambient", new BABYLON.Vector3(0, 0, -1), scene);
@@ -373,8 +378,6 @@ function Camera() {
         this.camera0 = new BABYLON.ArcRotateCamera("camera_model", 0, 0, 10, BABYLON.Vector3.Zero(), scene);
         this.camera1 = new BABYLON.ArcRotateCamera("camera_bake", 0, 0, 10, BABYLON.Vector3.Zero(), scene);
 
-        const fov = parseFloat(document.getElementById('input-camera-fov').value);
-
         this.camera0.setPosition(new BABYLON.Vector3(100, 100, 100));
         this.camera0.setTarget(BABYLON.Vector3.Zero());
         this.camera0.lowerRadiusLimit = 1;
@@ -387,7 +390,7 @@ function Camera() {
         this.camera0.inertia = 0.9; //def: 0.9
         this.camera0.minZ = 0.1;
         this.camera0.maxZ = 2000;
-        this.camera0.fov = fov; //def: 0.8
+        this.camera0.fov = parseFloat(document.getElementById('input-camera-fov').value); //def: 0.8
 
         this.camera1.setPosition(new BABYLON.Vector3(73, 73, 73));
         this.camera1.setTarget(new BABYLON.Vector3(13, 13, 13));
@@ -401,21 +404,20 @@ function Camera() {
         this.camera1.inertia = 0.9;
         this.camera1.minZ = 0.1;
         this.camera1.maxZ = 2000;
-        this.camera1.fov = fov;
+        this.camera1.fov = parseFloat(document.getElementById('input-camera-fov').value);
     }
 
     this.switchCamera = function() {
         scene.activeCamera.detachControl(canvas);
         if (MODE == 0) {
             scene.activeCamera = this.camera0;
-            this.camera0.attachControl(canvas, true);
         } else if (MODE == 1) {
-            scene.activeCamera = this.camera0;
-            this.camera0.attachControl(canvas, true);
+            scene.activeCamera = (document.getElementById('input-pt-source').value == 'model') ? this.camera0 : this.camera1;
         } else if (MODE == 2) {
             scene.activeCamera = this.camera1;
-            this.camera1.attachControl(canvas, true);
         }
+
+        scene.activeCamera.attachControl(canvas, true);
         ui.domCameraFov.value = scene.activeCamera.fov;
     }
 
@@ -429,7 +431,9 @@ function Camera() {
                 if (bakery.selected) { // zoom to selected mesh
                     this.setFramingBehavior(this.camera1, bakery.selected);
                 } else {
-                    this.setFramingBehavior(this.camera1, builder.mesh);
+                    const mesh = BABYLON.Mesh.MergeMeshes(bakery.meshes, false, true);
+                    this.setFramingBehavior(this.camera1, mesh);
+                    mesh.dispose();
                 }
             } else {
                 this.setFramingBehavior(this.camera1, builder.mesh);
@@ -447,13 +451,11 @@ function Camera() {
         if (!mesh) return;
         mesh.computeWorldMatrix(true);
         const bounds = mesh.getBoundingInfo();
-
-        const bbox = bounds.boundingBox;
-        const radiusWorld = bbox.maximumWorld.subtract(bbox.minimumWorld).scale(0.5);
-        const centerWorld = bbox.minimumWorld.add(radiusWorld);
-        const radius = (bounds.boundingSphere.radiusWorld + offset) / scene.activeCamera.fov;
-
-        return { radius: radius, target: centerWorld };
+        const radiusWorld = bounds.boundingBox.maximumWorld.subtract(bounds.boundingBox.minimumWorld).scale(0.5);
+        return {
+            radius: (bounds.boundingSphere.radiusWorld + offset) / scene.activeCamera.fov,
+            target: bounds.boundingBox.minimumWorld.add(radiusWorld)
+        };
     }
 
     this.toggleCameraAutoRotation = function() {
@@ -1240,13 +1242,13 @@ function Builder(scene) {
         const x = Math.round(scene.pointerX); // by @kikoshoung
         const y = engine.getRenderHeight() - Math.round(scene.pointerY);
         const pixels = readTexturePixels(engine._gl, renderPickTarget._texture._hardwareTexture.underlyingResource, x, y, 1, 1);
-        const colorId = `${pixels[0]}_${pixels[1]}_${pixels[2]}`;        
+        const colorId = `${pixels[0]}_${pixels[1]}_${pixels[2]}`;
         if (this.bufferWorld[this.rttColorsMap[colorId]])
             return this.rttColorsMap[colorId];
-        return -1;
+        return undefined;
     }
 
-    this.getIndexAtPosition = function(pos) { // return undefined
+    this.getIndexAtPosition = function(pos) {
         return this.positionsMap[`${pos.x}_${pos.y}_${pos.z}`];
     }
 
@@ -1258,14 +1260,14 @@ function Builder(scene) {
         const idx = this.getIndexAtPosition(pos);
         if (idx > -1)
             return this.voxels[idx];
-        return null;
+        return undefined;
     }
 
     this.getVoxelAtXYZ = function(x, y, z) {
         const idx = this.getIndexAtXYZ(x, y, z);
         if (idx > -1)
             return this.voxels[idx];
-        return null;
+        return undefined;
     }
 
     this.getVoxelsByPosition = function(pos) {
@@ -1281,6 +1283,13 @@ function Builder(scene) {
 
     this.getVoxelsByVisibility = function(isVisible) { // dup
         return this.voxels.filter(i => i.visible === isVisible);
+    }
+
+    this.getVoxelsVisibilityByColor = function(hex) {
+        const idx = this.findIndexByColor(hex);
+        if (idx > -1)
+            return this.voxels[idx].visible;
+        return false;
     }
 
     this.findIndexByPosition = function(pos) { // no dup
@@ -1335,6 +1344,14 @@ function Builder(scene) {
             this.voxels[i].visible = isVisible;
     }
 
+    this.setVoxelsVisibilityByColor = function(hex, isVisible) {
+        const voxels = this.getVoxelsByColor(hex);
+        for (let i = 0; i < voxels.length; i++)
+            this.voxels[voxels[i].idx].visible = isVisible;
+        this.create();
+        this.update();
+    }
+
     this.setAllColorsAndUpdate = async function(hex = currentColor) {
         if (!await ui.showConfirm('replace all colors?')) return;
         for (let i = 0; i < this.voxels.length; i++) {
@@ -1362,7 +1379,7 @@ function Builder(scene) {
         this.update();
     }
 
-    this.deleteColorAndUpdate = async function(hex) {
+    this.deleteColorAndUpdate = function(hex) {
         const group = this.getVoxelsByColor(hex);
         if (group.length == 0) return;
         
@@ -1677,13 +1694,6 @@ function Ghosts(scene) {
         return this.thin.getBoundingInfo().boundingSphere.centerWorld;
     }
 
-    this.getCenterFromArray = function(voxels) {
-        this.createThin(voxels);
-        const center = this.thin.getBoundingInfo().boundingSphere.centerWorld;
-        this.disposeThin();
-        return center;
-    }
-
     this.init();
 }
 
@@ -1827,28 +1837,29 @@ function Palette() {
     this.init = function() {
         canvasPalette.width = canvasPalette.clientWidth;
         canvasPalette.height = canvasPalette.clientHeight;
-                
+
         canvasPalette.addEventListener("pointerdown", (ev) => {
             const hex = getCanvasColor(ctx, ev.offsetX, ev.offsetY);
             if (hex) {
                 if (palette.uniqueColors.includes(hex)) {
-                    if (ev.button == 2) {
-                        palette.setColorVisible(hex, !palette.isColorVisible(hex));
-                    } else {
-                        currentColor = hex;
-                        uix.colorPicker.value = BABYLON.Color3.FromHexString(hex);
-                    }
+                    currentColor = hex;
+                    uix.colorPicker.value = BABYLON.Color3.FromHexString(hex);
                 }
             }
+        }, false);
+                
+        canvasPalette.addEventListener("contextmenu", (ev) => {
+            ev.preventDefault();
+            const hex = getCanvasColor(ctx, ev.offsetX, ev.offsetY);
+            if (hex && palette.uniqueColors.includes(hex))
+                builder.setVoxelsVisibilityByColor(hex, !builder.getVoxelsVisibilityByColor(hex));
         }, false);
 
         if (isMobile) {
             canvasPalette.addEventListener("dblclick", (ev) => {
                 const hex = getCanvasColor(ctx, ev.offsetX, ev.offsetY);
-                if (hex) {
-                    if (palette.uniqueColors.includes(hex))
-                        palette.setColorVisible(hex, !palette.isColorVisible(hex));
-                }
+                if (hex && palette.uniqueColors.includes(hex))
+                    builder.setVoxelsVisibilityByColor(hex, !builder.getVoxelsVisibilityByColor(hex));
             }, false);
         }
 
@@ -1888,28 +1899,13 @@ function Palette() {
     
     this.addColor = function(x, y, hex) {
         ctx.strokeStyle = 'transparent';
-        if (!this.isColorVisible(hex)) {
+        if (!builder.getVoxelsVisibilityByColor(hex)) {
             ctx.lineWidth = 2;
             ctx.strokeStyle = 'orange';
             ctx.strokeRect(x, y, W, H);
         }
         ctx.fillStyle = hex;
         ctx.fillRect(x, y, W, H);
-    }
-
-    this.setColorVisible = function(hex, isVisible) {
-        const voxels = builder.getVoxelsByColor(hex);
-        for (let i = 0; i < voxels.length; i++)
-            builder.voxels[voxels[i].idx].visible = isVisible;
-        builder.create();
-        builder.update();
-    }
-
-    this.isColorVisible = function(hex) {
-        const idx = builder.findIndexByColor(hex);
-        if (idx > -1)
-            return builder.voxels[ idx ].visible;
-        return false;
     }
 
     this.init();
@@ -2483,14 +2479,14 @@ function Tool(scene) {
             case 'bucket':
                 this.bucket(builder.voxels[index].color);
                 break;
-            case 'hidecolor':
-                palette.setColorVisible(builder.voxels[index].color, false);
+            case 'hide_color':
+                builder.setVoxelsVisibilityByColor(builder.voxels[index].color, false);
                 break;
-            case 'isolatecolor':
+            case 'isolate_color':
                 builder.setVoxelsVisibility(false);
-                palette.setColorVisible(builder.voxels[index].color, true);
+                builder.setVoxelsVisibilityByColor(builder.voxels[index].color, true);
                 break;
-            case 'deletecolor':
+            case 'delete_color':
                 builder.deleteColorAndUpdate(builder.voxels[index].color);
                 break;
             case 'box_add':
@@ -2811,7 +2807,7 @@ function Tool(scene) {
 
     this.setPickInfo = function(pick, onHit) {
         const index = builder.getIndexAtPointer();
-        if (index > -1) {
+        if (index) {
             const norm = normalProbe(builder.voxels[index].position);
             if (norm) {
                 pick.INDEX = index;
@@ -2839,15 +2835,15 @@ function Tool(scene) {
     this.toolSelector = function(toolName, finishTransforms = false) {
         this.name = toolName;
 
-        const elem = document.getElementsByClassName('tool_' + this.name);
+        const elems = document.getElementsByClassName('tool_' + this.name);
         for (const i of document.querySelectorAll('li'))
             if (i.classList.contains("tool_selector"))
                 i.classList.remove("tool_selector");
         for (const i of document.querySelectorAll('button'))
             if (i.classList.contains("tool_selector"))
                 i.classList.remove("tool_selector");
-        for (let i = 0; i < elem.length; i++)
-            elem[i].classList.add("tool_selector");
+        for (let i = 0; i < elems.length; i++)
+            elems[i].classList.add("tool_selector");
 
         scene.activeCamera.attachControl(canvas, true);
         helper.clearOverlays();
@@ -2859,6 +2855,91 @@ function Tool(scene) {
     }
 
     this.init();
+}
+
+
+// -------------------------------------------------------
+// Tool Bakery
+
+
+function ToolBakery(scene) {
+    this.name = 'select';
+    this.selected = [];
+    this.pick = null;
+
+    this.onToolDown = function() {
+        this.pick = scene.pick(scene.pointerX, scene.pointerY, (mesh) => {
+            return bakery.meshes.includes(mesh);
+        });
+
+        if (this.pick && this.pick.hit) {
+            switch (this.name) {
+                case 'select':
+                    bakery.deselectMesh();
+                    uix.bindTransformGizmo(this.pick.pickedMesh);
+                    break;
+                case 'merge':
+                    const idx = this.selected.indexOf(this.pick.pickedMesh);
+                    if (idx == -1) {
+                        this.selected.push(this.pick.pickedMesh);
+                        highlightOverlayMesh(this.pick.pickedMesh, COL_ORANGE_RGB);
+                        highlightOutlineMesh(this.pick.pickedMesh, COL_ORANGE_RGB);
+                    } else {
+                        this.selected.splice(idx, 1);
+                        this.pick.pickedMesh.renderOverlay = false;
+                        this.pick.pickedMesh.renderOutline = false;
+                    }
+                    break;
+            }
+        }
+    }
+
+    this.mergeBakes = function() {
+        if (this.selected.length == 1) {
+            ui.notification('pick more meshes ...');
+            return;
+        }
+        
+        if (this.selected.length > 1) {
+            for (let i = 0; i < this.selected.length; i++) {
+                this.selected[i].renderOverlay = false;
+                this.selected[i].renderOutline = false;
+            }
+
+            bakery.mergeSelected(this.selected);
+
+            this.selected = [];
+            this.toolSelector('select');
+        } else {
+            ui.notification('pick meshes to merge');
+        }
+    }
+
+    this.cancelSelection = function() {
+        for (let i = 0; i < this.selected.length; i++) {
+            this.selected[i].renderOverlay = false;
+            this.selected[i].renderOutline = false;
+        }
+        this.selected = [];
+        this.toolSelector('select');
+    }
+
+    this.toolSelector = function(toolName) {
+        this.name = toolName;
+
+        const elems = document.getElementsByClassName('tool_' + this.name);
+        for (let i of document.querySelectorAll('li'))
+            if (i.classList.contains("tool_bakery_selector"))
+                i.classList.remove("tool_bakery_selector");
+        for (let i of document.querySelectorAll('button'))
+            if (i.classList.contains("tool_bakery_selector"))
+                i.classList.remove("tool_bakery_selector");
+        for (let i = 0; i < elems.length; i++)
+            elems[i].classList.add("tool_bakery_selector");
+
+        bakery.deselectMesh();
+        uix.unbindTransformGizmo();
+    }
 }
 
 
@@ -3069,13 +3150,12 @@ function Voxelizer(scene) {
             if (isClear && !await ui.showConfirm('clear and replace all voxels?')) return;
 
             const data = window.voxelizeBake(bakery.selected);
+            ui.setMode(0);
 
             if (isClear) {
-                ui.setMode(0);
                 builder.setDataFromArray(data);
                 builder.normalizeVoxelPositions();
             } else {
-                ui.setMode(0);
                 xformer.beginNewObject(data, false);
             }
             clearScene();
@@ -3459,14 +3539,14 @@ function Bakery(scene) {
 
     this.exportOptions = {
         shouldExportNode: (node) => {
-            return bakery.meshes.includes(node);
+            return this.meshes.includes(node);
         }
     };
 
     this.exportOptionsSelected = {
         shouldExportNode: (node) => {
-            if (ui.domExportSelectedBake.checked && bakery.selected)
-                return bakery.selected === node;
+            if (ui.domExportSelectedBake.checked && this.selected)
+                return this.selected === node;
             return false;
         }
     };
@@ -3602,12 +3682,12 @@ function Bakery(scene) {
 
             const mesh = BABYLON.Mesh.MergeMeshes(this.meshes, true, true);
             resetPivot(mesh);
-            mesh.name = '_merge';
+            mesh.name = '_merged';
 
-            clearMeshArray();
+            this.clearMeshArray();
 
             material.setPBRTexture();
-            mesh.material = material.mat_pbr.clone('_merge');
+            mesh.material = material.mat_pbr.clone('_merged');
 
             mesh.checkCollisions = true;
             mesh.receiveShadows = true;
@@ -3617,9 +3697,39 @@ function Bakery(scene) {
             uix.bindTransformGizmo(mesh);
             uix.gizmo.attachToMesh(mesh);
             
-            bakery.meshes.push(mesh);
-            bakery.createBakeryList();
+            this.meshes.push(mesh);
+            this.createBakeryList();
         }
+    }
+
+    this.mergeSelected = function(bakes) {
+        const mesh = BABYLON.Mesh.MergeMeshes(bakes, true, true);
+        resetPivot(mesh);
+        mesh.name = '_merged';
+
+        for (let i = 0; i < bakes.length; i++) {
+            this.meshes.splice(this.meshes.indexOf(bakes[i]), 1);
+            if (bakes[i].material) {
+                if (bakes[i].material.albedoTexture)
+                    bakes[i].material.albedoTexture.dispose();
+                bakes[i].material.dispose();
+            }
+            bakes[i].dispose();
+        }
+
+        material.setPBRTexture();
+        mesh.material = material.mat_pbr.clone('_merged');
+
+        mesh.checkCollisions = true;
+        mesh.receiveShadows = true;
+        light.addMesh(mesh);
+        light.updateShadowMap();
+
+        uix.bindTransformGizmo(mesh);
+        uix.gizmo.attachToMesh(mesh);
+        
+        this.meshes.push(mesh);
+        this.createBakeryList();
     }
 
     this.deleteSelected = async function() {
@@ -3657,9 +3767,9 @@ function Bakery(scene) {
     }
 
     this.onGizmoAttached = function(mesh) {
-        bakery.deselectMesh(); // on user select
-        bakery.selectMesh(mesh);
-        bakery.getMaterial();
+        this.deselectMesh(); // on user select
+        this.selectMesh(mesh);
+        this.getMaterial();
     }
 
     this.setBakesVisibility = function(isVisible) {
@@ -3679,7 +3789,7 @@ function Bakery(scene) {
     this.clearBakes = async function(isAlert = false) {
         if (this.meshes.length > 0) {
             if (isAlert && !await ui.showConfirm('clear all baked meshes?')) return;
-            clearMeshArray();
+            this.clearMeshArray();
             this.selected = null;
             this.createBakeryList();
             uix.unbindTransformGizmo();
@@ -3758,25 +3868,25 @@ function Bakery(scene) {
 
     this.createBakeryList = function() {
         ui.domBakeryList.innerHTML = "";
-        for (let i = 0; i < bakery.meshes.length; i++) {
+        for (let i = 0; i < this.meshes.length; i++) {
             const item = document.createElement('div');
             const name = document.createElement('div');
             name.classList.add('item_name');
-            name.innerHTML = bakery.meshes[i].name;
+            name.innerHTML = this.meshes[i].name;
             name.spellcheck = false;
             item.onclick = () => {
-                bakery.deselectMesh();
-                bakery.selectMesh(bakery.meshes[i]);
+                this.deselectMesh();
+                this.selectMesh(this.meshes[i]);
                 uix.bindTransformGizmo(this.meshes[i]);
                 uix.gizmo.attachToMesh(this.meshes[i]);
             };
             item.onkeyup = () => {
-                bakery.meshes[i].name = name.innerHTML;
+                this.meshes[i].name = name.innerHTML;
             };
             item.onpaste = (ev) => {
                 ev.preventDefault();
                 name.innerHTML = ev.clipboardData.getData('Text');
-                bakery.meshes[i].name = name.innerHTML;
+                this.meshes[i].name = name.innerHTML;
             };
             item.ondblclick = () => {
                 name.contentEditable = true;
@@ -3793,8 +3903,8 @@ function Bakery(scene) {
 
     this.bakeListSelect = function(mesh) {
         let idx = -1;
-        for (let i = 0; i < bakery.meshes.length; i++)
-            if (bakery.meshes[i] === mesh)
+        for (let i = 0; i < this.meshes.length; i++)
+            if (this.meshes[i] === mesh)
                 idx = i;
         
         if (ui.domBakeryList.children[idx]) {
@@ -3818,7 +3928,7 @@ function Bakery(scene) {
                     if (container.meshes[i].name !== '__root__') {
                         const baked = container.meshes[i].clone(container.meshes[i].name);
                         normalizeMeshGLTF(baked, container.meshes[i].material.clone(container.meshes[i].name));
-                        bakery.meshes.push(baked);
+                        this.meshes.push(baked);
                         count += 1;
                     }
                 }
@@ -3828,12 +3938,15 @@ function Bakery(scene) {
                     ui.notification('unable to load baked meshes');
                 } else {
                     if (!isInsideBakery) {
-                        (isLoadBakery) ? ui.setMode(2) : ui.setMode(0);
-                    } else {
-                        builder.setMeshVisibility(false); // user may act faster
+                        if (isLoadBakery) {
+                            ui.setMode(2);
+                        } else {
+                            ui.setMode(0);
+                            bakery.setBakesVisibility(false);
+                        }
                     }
-                    bakery.updateReflectionTextures();
-                    bakery.createBakeryList();
+                    this.updateReflectionTextures();
+                    this.createBakeryList();
                     light.updateShadowMap();
                 }
                 engine.hideLoadingUI();
@@ -3842,6 +3955,20 @@ function Bakery(scene) {
                 ui.notification("unable to load bake");
                 console.error(reason.message);
             });
+    }
+
+    this.clearMeshArray = function () {
+        scene.blockfreeActiveMeshesAndRenderingGroups = true; // save unnecessary
+        for (let i = 0; i < this.meshes.length; i++) { // dispose() computation
+            if (this.meshes[i].material.albedoTexture)
+                this.meshes[i].material.albedoTexture.dispose();
+            if (this.meshes[i].material.reflectionTexture)
+                this.meshes[i].material.reflectionTexture.dispose();
+            this.meshes[i].material.dispose();
+            this.meshes[i].dispose();
+        }
+        scene.blockfreeActiveMeshesAndRenderingGroups = false;
+        this.meshes = [];
     }
 
     function normalizeMeshGLTF(mesh, material) {
@@ -3857,20 +3984,6 @@ function Bakery(scene) {
         mesh.checkCollisions = true;
         mesh.receiveShadows = true;
         light.addMesh(mesh);
-    }
-
-    function clearMeshArray() {
-        scene.blockfreeActiveMeshesAndRenderingGroups = true; // save unnecessary
-        for (let i = 0; i < bakery.meshes.length; i++) { // dispose() computation
-            if (bakery.meshes[i].material.albedoTexture)
-                bakery.meshes[i].material.albedoTexture.dispose();
-            if (bakery.meshes[i].material.reflectionTexture)
-                bakery.meshes[i].material.reflectionTexture.dispose();
-            bakery.meshes[i].material.dispose();
-            bakery.meshes[i].dispose();
-        }
-        scene.blockfreeActiveMeshesAndRenderingGroups = false;
-        bakery.meshes = [];
     }
 }
 
@@ -3926,8 +4039,9 @@ function Snapshot(scene) {
             return;
         }
         if (data !== '0') {
+            ui.setMode(2);
             bakery.clearBakes(false);
-            bakery.loadBakes(data, true, true);
+            bakery.loadBakes(data, false, true);
         }
     }
 
@@ -4032,7 +4146,7 @@ function Memory() {
 function Project(scene) {
     function serializeScene(voxels, meshes) {
         const json = {
-            version: "Voxel Builder 4.3.0",
+            version: "Voxel Builder 4.3.1",
             project: {
                 name: "name",
                 voxels: builder.voxels.length,
@@ -4094,19 +4208,8 @@ function Project(scene) {
     }
 
     this.load = function(data) {
-        if (data.startsWith(';')) {
-            loader(parseINI(data)); // backward compatibity < 4.0.3
-            return;
-        }
+        data = JSON.parse(data);
         
-        try {
-            loader(JSON.parse(data));
-        } catch(err) {
-            ui.notification("unsupported data format");
-        }
-    }
-
-    function loader(data) {
         // project
         ui.domProjectName.value = data.project.name;
 
@@ -4115,7 +4218,6 @@ function Project(scene) {
             updateScene(data.scene)
 
         // data.voxels
-        ui.setMode(0);
         builder.setDataFromString(data.data.voxels);
         clearSceneAndReset();
 
@@ -4153,7 +4255,6 @@ function Project(scene) {
     this.importBakes = function(data) {
         data = JSON.parse(data);
         if (data.data.meshes) {
-            ui.setMode(2);
             bakery.loadBakes(data.data.meshes, true);
         } else {
             ui.notification('no baked meshes');
@@ -4212,9 +4313,7 @@ function Project(scene) {
         }
         engine.displayLoadingUI();
 
-        let exports;
-
-        exports = bakery.exportOptions;
+        let exports = bakery.exportOptions;
         if (ui.domExportSelectedBake.checked && bakery.selected)
             exports = bakery.exportOptionsSelected;
 
@@ -4269,7 +4368,6 @@ function Project(scene) {
                     visible: true
                 });
             }
-            ui.setMode(0);
             builder.setDataFromArray(voxels, true);
             builder.normalizeVoxelPositions(false);
             clearSceneAndReset();
@@ -4373,6 +4471,7 @@ function UserInterface(scene) {
     this.domInfo = document.getElementById('info').children;
     this.domInfoParent = document.getElementById('info');
     this.domInfoTool = document.getElementById('info_tool');
+    this.domInfoRender = document.getElementById('info_render');
     this.domProgressBar = document.getElementById('progressbar');
     this.domLoadingScreen = document.getElementById('loadingscreen');
     const styleMenuR = '-200px';
@@ -4478,6 +4577,7 @@ function UserInterface(scene) {
             builder.setMeshVisibility(false);
             bakery.setBakesVisibility(true);
             bakery.createBakeryList();
+            toolBakery.toolSelector('select');
             helper.displayGridPlane(helper.isFloorPlaneActive, true);
             helper.displayWorkplane(helper.isWorkplaneActive);
             helper.toggleAxisPlane(false);
@@ -4550,7 +4650,6 @@ function UserInterface(scene) {
             this.domToolbarR.children[5].style.display = 'none'; // MODEL
             this.domToolbarR.children[6].style.display = 'none'; // PAINT
             this.domToolbarR.children[7].style.display = 'none'; // VOXELS
-            this.domToolbarL.children[1].style.display = 'none'; // RENDER
             this.domToolbarL.children[6].style.display = 'none'; // STORAGE
             this.domToolbarL.children[7].style.display = 'none'; // GROUP
         }
@@ -4630,7 +4729,7 @@ function UserInterface(scene) {
             arrL = [ 3, 4, 5, 6 ];
             arrR = [ 2, 3, 4, 5, 6, 7, 8, 9, 10 ];
         } else if (mode == 2) {
-            arrL = [ 0, 5, 6 ];
+            arrL = [ 5, 6 ];
             arrR = [ 3, 4, 5, 6, 7, 8 ];
         }
 
@@ -4695,8 +4794,6 @@ function UserInterface(scene) {
         if (MODE == 0) {
             this.domInfo[1].innerHTML = builder.voxels.length + ' VOX';
             this.domInfo[2].innerHTML = palette.uniqueColors.length + ' COL';
-        } else if (MODE == 1) {
-            //
         } else if (MODE == 2) {
             this.domInfo[1].innerHTML = bakery.meshes.length + ' MSH';
             this.domInfo[2].innerHTML = scene.getTotalVertices() + ' VTX';
@@ -4880,15 +4977,15 @@ function UserInterface(scene) {
 
 
 function UserInterfaceAdvanced(scene) {
-    this.advancedTexture = null;
-    this.utilLayer = null;
+    this.advancedTexture = undefined;
+    this.utilLayer = undefined;
     this.colorPicker = undefined;
-    this.gizmo = null;
-    this.gizmoVoxel = null;
-    this.sunNode = null;
-    this.sunGizmoUp = null;
-    this.sunGizmoNews = null;
-    this.isSunLocatorActive = null;
+    this.gizmo = undefined;
+    this.gizmoVoxel = undefined;
+    this.sunNode = undefined;
+    this.sunGizmoUp = undefined;
+    this.sunGizmoNews = undefined;
+    this.isSunLocatorActive = undefined;
 
     this.init = function() {
         this.advancedTexture = BABYLON.GUI.AdvancedDynamicTexture.CreateFullscreenUI("UI", {}, scene);
@@ -4950,7 +5047,7 @@ function UserInterfaceAdvanced(scene) {
     this.unbindVoxelGizmo = function() {
         if (this.gizmoVoxel)
             this.gizmoVoxel.dispose();
-        this.gizmoVoxel = null;
+        this.gizmoVoxel = undefined;
     }
 
     this.bindTransformGizmo = function(meshes) {
@@ -4993,7 +5090,10 @@ function UserInterfaceAdvanced(scene) {
                 });
             });
 
-        this.gizmo.attachableMeshes = meshes;
+        (meshes.length > 1) ?
+            this.gizmo.attachableMeshes = meshes :
+            this.gizmo.attachableMeshes = [meshes];
+
         this.gizmo.onAttachedToMeshObservable.add((mesh) => {
             (mesh) ? bakery.onGizmoAttached(mesh) : bakery.deselectMesh();
         });
@@ -5002,7 +5102,7 @@ function UserInterfaceAdvanced(scene) {
     this.unbindTransformGizmo = function() {
         if (this.gizmo)
             this.gizmo.dispose();
-        this.gizmo = null;
+        this.gizmo = undefined;
     }
 
     this.initLightLocator = function() {
@@ -5250,7 +5350,7 @@ scene.onPointerObservable.add((pInfo) => {
         case BABYLON.PointerEventTypes.POINTERDOWN:
             if (pInfo.event.button > 0) return;
             if (MODE == 0) tool.handleToolDown(pInfo.pickInfo);
-            else if (MODE == 2) uix.bindTransformGizmo(bakery.meshes);
+            if (MODE == 2) toolBakery.onToolDown();
             break;
         case BABYLON.PointerEventTypes.POINTERMOVE:
             if (MODE == 0) tool.handleToolMove(pInfo.pickInfo);
@@ -5280,8 +5380,7 @@ document.addEventListener("keydown", (ev) => {
     switch (ev.key.toLowerCase()) {
         case 'enter':
             if (ev.target instanceof HTMLButtonElement) return;
-            if (MODE == 0)
-                xformer.apply();
+            if (MODE == 0) xformer.apply();
             break;
         case ' ':
             if (MODE == 0) tool.toolSelector('camera', true);
@@ -5327,10 +5426,14 @@ document.addEventListener("keydown", (ev) => {
                 window.pt.frameCamera() : camera.frame();
             break;
         case 'o':
-            camera.switchOrtho();
+            if (MODE != 1) camera.switchOrtho();
             break;
         case 'r':
-            (window.pt && window.pt.isActive()) ? ui.setMode(0) : ui.setMode(1);
+            if (window.pt && window.pt.isActive()) {
+                (window.pt.source.value == 'model') ? ui.setMode(0) : ui.setMode(2);
+            } else {
+                ui.setMode(1);
+            }
             break;
     }
 
@@ -5339,19 +5442,19 @@ document.addEventListener("keydown", (ev) => {
     
     if (MODE == 0) {
         if (ev.ctrlKey && ev.key.toLowerCase() === 'z') {
-            xformer.apply();
             ev.preventDefault();
+            xformer.apply();
             memory.undo();
         }
         if (ev.ctrlKey && ev.key.toLowerCase() === 'x') {
-            xformer.apply();
             ev.preventDefault();
+            xformer.apply();
             memory.redo();
         }
     }
 }, false);
 
-document.addEventListener("keyup", (ev) => {
+document.addEventListener("keyup", () => {
     if (tool.last) {
         tool.toolSelector(tool.last);
         tool.last = null;
@@ -5363,11 +5466,11 @@ function fileHandler(file) {
     const url = URL.createObjectURL(file);
     const reader = new FileReader();
     reader.onload = () => {
-        if (ext === 'json' || ext === 'vbx') project.load(reader.result);
-        if (ext === 'obj') if (MODE == 0) voxelizer.importMeshOBJ(url);
-        if (ext === 'glb') if (MODE == 0) voxelizer.importMeshGLB(url);
-        if (ext === 'vox') project.loadMagicaVoxel(reader.result);
-        if (ext === 'hdr') hdri.loadHDR(url);
+        if (ext == 'json') project.load(reader.result);
+        if (ext == 'obj') if (MODE == 0) voxelizer.importMeshOBJ(url);
+        if (ext == 'glb') if (MODE == 0) voxelizer.importMeshGLB(url);
+        if (ext == 'vox') project.loadMagicaVoxel(reader.result);
+        if (ext == 'hdr') hdri.loadHDR(url);
         if (MODE == 0) {
             if (['jpg','png','svg'].includes(ext)) voxelizer.voxelize2D(reader.result);
         } else if (MODE == 2) {
@@ -5375,9 +5478,9 @@ function fileHandler(file) {
         }
         URL.revokeObjectURL(url);
     }
-    if (ext === 'json' || ext === 'vbx') {
+    if (ext == 'json') {
         reader.readAsText(file);
-    } else if (ext === 'vox') {
+    } else if (ext == 'vox') {
         reader.readAsArrayBuffer(file);
     } else {
         reader.readAsDataURL(file);
@@ -5388,13 +5491,12 @@ function fileHandlerNoDrop(file, type) {
     const url = URL.createObjectURL(file);
     const reader = new FileReader();
     reader.onload = () => {
-        if (type === 'import_voxels') project.importVoxels(reader.result);
-        if (type === 'import_bakes')  project.importBakes(reader.result);
+        if (type == 'import_voxels') project.importVoxels(reader.result);
+        if (type == 'import_bakes')  project.importBakes(reader.result);
         URL.revokeObjectURL(url);
     }
-    if (type === 'import_voxels' || type === 'import_bakes') {
+    if (type == 'import_voxels' || type == 'import_bakes')
         reader.readAsText(file);
-    }
 }
 
 function dropHandler(ev) {
@@ -5457,11 +5559,8 @@ window.addEventListener("resize", () => {
 
     updateAxisViewViewport();
 
-    if (MODE == 0)
-        palette.create();
-
-    if (MODE == 2)
-        bakery.createBakeryList();
+    if (MODE == 0) palette.create();
+    if (MODE == 2) bakery.createBakeryList();
 
     if (isOffScreen(ui.domHover))
         ui.hoverTranslate(ui.domHover, 0, 0);
@@ -5751,9 +5850,9 @@ function downloadBlob(blob, filename) {
 }
 
 async function saveFile(blob, filename, reject) {
-    if (!isElectron() && !isMobile) { // to show a file dialog in Chrome
+    if ("showSaveFilePicker" in window) {
         try {
-            const fileHandle = await self.showSaveFilePicker({
+            const fileHandle = await window.showSaveFilePicker({
                 suggestedName: filename,
                 types: [ { description: "File" } ]
             });
@@ -5768,7 +5867,7 @@ async function saveFile(blob, filename, reject) {
         } catch (err) {
             // canceled
         }
-    } else { // showSaveFilePicker browser compatibility
+    } else {
         reject();
     }
 }
@@ -5792,11 +5891,9 @@ function dataURItoBlob(dataURI) {
     const mimeString = dataURI.split(',')[0].split(':')[1].split(';')[0]
     const ab = new ArrayBuffer(byteString.length);
     const ia = new Uint8Array(ab);
-    for (let i = 0; i < byteString.length; i++) {
+    for (let i = 0; i < byteString.length; i++)
         ia[i] = byteString.charCodeAt(i);
-    }
-    const blob = new Blob([ ab ], { type: mimeString });
-    return blob;
+    return new Blob([ ab ], { type: mimeString });
 }
 
 function toggleFullscreen() {
@@ -5831,12 +5928,6 @@ function aspectRatioFit(srcW, srcH, maxW, maxH) {
 
 function getStyleRoot(key) {
     return getComputedStyle(document.querySelector(':root')).getPropertyValue(key);
-}
-
-function evenNumberIndicatorInt(elem) {
-    (Math.abs(parseInt(elem.value) % 2) == 1) ? // detect odd numbers
-        elem.style.color = 'indianred' :
-        elem.style.color = getStyleRoot('--input-color');
 }
 
 function rgbIntToHex(r, g, b) {
@@ -5913,34 +6004,4 @@ function timeFormat(t) {
     if (m < 10) { m = "0" + m; }
     if (s < 10) { s = "0" + s; }
     return h + ':' + m + ':' + s;
-}
-
-function parseINI(data) {
-    const regex = {
-        section: /^\s*\[\s*([^\]]*)\s*\]\s*$/,
-        param: /^\s*([^=]+?)\s*=\s*(.*?)\s*$/,
-        comment: /^\s*;.*$/
-    };
-    const value = {};
-    const lines = data.split(/[\r\n]+/);
-    let section = null;
-    lines.forEach((line) => {
-        if (regex.comment.test(line)) {
-            return;
-        } else if (regex.param.test(line)) {
-            const match = line.match(regex.param);
-            if (section) {
-                value[section][match[1]] = match[2];
-            } else{
-                value[match[1]] = match[2];
-            }
-        } else if (regex.section.test(line)) {
-            const match = line.match(regex.section);
-            value[match[1]] = {};
-            section = match[1];
-        } else if (line.length == 0 && section) {
-            section = null;
-        }
-    });
-    return value;
 }